requirements.txt
aiogram>=3.3.0,<4.0
SQLAlchemy>=2.0.30,<2.1
psycopg2-binary>=2.9.0,<3.0
pydantic>=2.6.0,<3.0
pydantic-settings>=2.5.0,<3.0
redis>=5.0.1,<6.0
python-dotenv>=1.0.1,<2.0
uvicorn[standard]>=0.30.0,<1.0
fastapi>=0.112.0,<1.0


